use the gentle version of the minifier as the default
Remove the aggressive minifier, and remove the utility that merges reprozip files. There has been no apparent need for merging reprozip pack files. The pack file spec is also subject to change.
