--- neurodocker/reprozip/__init__.py
@@ -1,2 +1,0 @@
-from neurodocker.reprozip.trace import ReproZipMinimizer
-from neurodocker.reprozip.merge import merge_pack_files
